refactor(collapse): separate component
