+++ src/components/Collapse/index.ts
@@ -1,0 +1,1 @@
+export { Collapse } from './Collapse';
