+++ src/components/Layout/NavigationGroup.tsx
@@ -1,71 +1,11 @@
-import React, {
-  ReactElement,
-  ReactNode,
-  useEffect,
-  useRef,
-  useState,
-} from 'react';
+import React, { ReactElement, ReactNode } from 'react';
 
 import styled, { css } from 'styled-components';
 
 import { StyledProps } from 'types/StyledProps';
 
-/**
- * START ======== Collapse Component ========
- * */
-const StyledCollapseWrapper = styled.div<StyledCollapseWrapperProps>`
-  transition: height 0.3s cubic-bezier(0.4, 0, 0.2, 1) 0ms;
-  overflow: hidden;
-  position: relative;
-  min-height: ${({ $collapseSize }) =>
-    $collapseSize ? `${$collapseSize}px` : '0px'};
-  height: ${({ $height }) => `${$height}px`};
-`;
+import { Collapse } from '../Collapse';
 
-type StyledCollapseWrapperProps = StyledProps<
-  CollapseWrapperProps & { height: number }
->;
-
-export type CollapseWrapperProps = {
-  className?: string;
-  isExpand: boolean;
-  collapseSize?: number;
-  children: ReactNode;
-};
-export const CollapseWrapper = ({
-  className,
-  isExpand,
-  collapseSize,
-  children,
-}: CollapseWrapperProps) => {
-  const [height, setHeight] = useState(0);
-  const ref = useRef<HTMLDivElement | null>(null);
-  useEffect(() => {
-    if (ref.current && isExpand) {
-      setHeight(ref.current.getBoundingClientRect().height);
-    } else {
-      setHeight(0);
-    }
-  }, [isExpand, setHeight]);
-  return (
-    <StyledCollapseWrapper
-      className={className}
-      $height={height}
-      $isExpand={isExpand}
-      $collapseSize={collapseSize || 0}
-    >
-      <div ref={ref}>{children}</div>
-    </StyledCollapseWrapper>
-  );
-};
-
-/**
- * END ======== Collapse Component ========
- * */
-
-/**
- * START ======== Navigation Group Component ========
- * */
 const Wrapper = styled.div`
   display: flex;
   flex-direction: column;
@@ -104,7 +44,7 @@ const StyledNavigationItem = styled.div<StyledProps<NavigationItemProps>>`
 `;
 
 const StyledItemWrapper = styled.div``;
-const StyledGroupItemWrapper = styled(CollapseWrapper)`
+const StyledGroupItemWrapper = styled(Collapse)`
   padding-left: var(--amino-space);
   margin-left: var(--amino-space);
   border-left: 1px solid var(--amino-gray-200);
@@ -159,7 +99,3 @@ export const NavigationGroup = ({
     </Wrapper>
   );
 };
-
-/**
- * END ======== Navigation Group Component ========
- * */
